+++ src/main/java/com/meoguri/linkocean/domain/tag/service/TagService.java
@@ -2,8 +2,8 @@ package com.meoguri.linkocean.domain.tag.service;
 
 import java.util.List;
 
+import com.meoguri.linkocean.domain.bookmark.entity.vo.Tags;
 import com.meoguri.linkocean.domain.profile.query.service.dto.GetProfileTagsResult;
-import com.meoguri.linkocean.domain.tag.entity.Tags;
 
 public interface TagService {
