+++ src/test/java/com/meoguri/linkocean/domain/bookmark/entity/BookmarkTest.java
@@ -16,10 +16,10 @@ import org.springframework.test.util.ReflectionTestUtils;
 
 import com.meoguri.linkocean.domain.bookmark.entity.vo.Category;
 import com.meoguri.linkocean.domain.bookmark.entity.vo.OpenType;
+import com.meoguri.linkocean.domain.bookmark.entity.vo.Tags;
 import com.meoguri.linkocean.domain.linkmetadata.entity.LinkMetadata;
 import com.meoguri.linkocean.domain.profile.command.entity.Profile;
 import com.meoguri.linkocean.domain.tag.entity.Tag;
-import com.meoguri.linkocean.domain.tag.entity.Tags;
 
 class BookmarkTest {
